renamed DefaultColorPicker to ColorPicker.svelte
