+++ src/components/controls/ColorPicker.svelte
